Made some changes

+++ index.php
@@ -39,6 +39,7 @@ $syno = new Synology();
 
 $path = "/test";
 if (isset($_POST['open']) && $_POST['open'] != null) $path = $_POST['open'];
+$backUrl = substr($path, 0, strrpos($path, "/"));
 ?>
 <!DOCTYPE html>
 <html>
@@ -100,16 +101,11 @@ if (isset($_POST['open']) && $_POST['open'] != null) $path = $_POST['open'];
                     <th></th>
                 </tr>
                 <?php
-                $list = $syno->GetList( array('method' => 'list', 'version' => 2, 'sort_by' => 'type', 'additional' => '["real_path", "size", "owner", "time", "perm", "mount_point_type", "type"]'), $path)->data->files;
-                //print_r(json_encode($list));
-
-                $backUrl = substr($path, 0, strrpos($path, "/"));
-
                 if ($path != "/test") echo "<tr><td><a style='cursor: pointer;' onclick=\"openFolder('$backUrl');\"><i class='fas fa-arrow-left fa-lg'></i></a></td><td></td><td></td><td></td><td></td><td></td><td></td><td></td></tr>";
 
-                foreach ($list as $item) {
-                    $downloadUrl = $syno->download('entry.cgi', 'SYNO.FileStation.Download', array('method' => 'download', 'version' => 2, 'mode' => '"download"'), $item->path);
+                $list = $syno->GetList( array('method' => 'list', 'version' => 2, 'sort_by' => 'type', 'additional' => '["real_path", "size", "owner", "time", "perm", "mount_point_type", "type"]'), $path)->data->files;
 
+                foreach ($list as $item) {
                     if (substr($item->name, 0, 1) !== "#") {
                         $owner = $item->additional->owner->user;
                         $info = json_encode($item);
@@ -122,18 +118,7 @@ if (isset($_POST['open']) && $_POST['open'] != null) $path = $_POST['open'];
                                     <td><i class='fas fa-folder-open fa-lg'></i></td>
                                     <td><a href='#' onclick=\"openFolder('$item->path');\">$item->name</a></td>
                                     <td>$type</td>
-                                    <td></td>
-                                    <td>$owner</td>
-                                    <td>
-                                        <form action='info.php' method='post'>
-                                            <input type='hidden' name='item' value='$info'/>
-                                            <button type='submit' class='btn btn-link' style='padding:0px;'>Info</button>
-                                        </form>
-                                    </td>
-                                    <td><a href='$downloadUrl'>Download</a></td>
-                                    <td><a href='delete.php?path=$item->path'>Delete</a></td>
-                                    <td><a href='#'>Link</a></td>
-                                </tr>";
+                                    <td></td>";
                         } else {
                             $type = "bestand";
                             $openInOffice = false;
@@ -156,19 +141,21 @@ if (isset($_POST['open']) && $_POST['open'] != null) $path = $_POST['open'];
                                     <td><i class='fas fa-file fa-lg'></i></td>
                                     <td><a href='$openUrl' onclick='edit(\"$path\", \"$item->name\", this)'>$item->name</a></td>
                                     <td>$type</td>
-                                    <td>$size</td>
-                                    <td>$owner</td>
-                                    <td>
-                                        <form action='info.php' method='post'>
-                                            <input type='hidden' name='item' value='$info'/>
-                                            <button type='submit' class='btn btn-link' style='padding:0px;'>Info</button>
-                                        </form>
-                                    </td>
-                                    <td><a href='$downloadUrl'>Download</a></td>
-                                    <td><a href='delete.php?path=$item->path'>Delete</a></td>
-                                    <td><a href='#'>Link</a></td>
-                                </tr>";
+                                    <td>$size</td>";
                         }
+
+                        $downloadUrl = $syno->download('entry.cgi', 'SYNO.FileStation.Download', array('method' => 'download', 'version' => 2, 'mode' => '"download"'), $item->path);
+
+                        echo "  <td>$owner</td>
+                                <td>
+                                    <form action='info.php' method='post'>
+                                        <input type='hidden' name='item' value='$info'/>
+                                        <button type='submit' class='btn btn-link' style='padding:0px;'>Info</button>
+                                    </form>
+                                </td>
+                                <td><a href='$downloadUrl'>Download</a></td>
+                                <td><a href='delete.php?path=$item->path' onclick=\"return  confirm('Weet je zeker dat je $item->name wilt verwijderen?')\">Delete</a></td>
+                                <td><a href='#'>Link</a></td></tr>";
                     }
                 }
                 ?>
@@ -213,21 +200,45 @@ if (isset($_POST['open']) && $_POST['open'] != null) $path = $_POST['open'];
             }
 
             function edit(path, filename, event) {
+                var filePath = path + "/" + filename;
+
+                //ITHit.WebDAV.Client.DocManager.DavProtocolEditDocument(
+                    //filePath,                                 // Document URL(s)
+                    //path,                                     // Mount URL
+                    //errorCallback(),                          // Function to call if protocol app is not installed
+                    //null,                                     // Reserved
+                    //'Current',                                // Which browser to copy cookies from: 'Current', 'All', 'None'
+                    //'.AspNet.ApplicationCookie',              // Cookie(s) to copy.
+                    //'/',                                      // URL to navigate to if any cookie from the list is not found.
+                    //'Edit'                                    // Command to execute: 'Edit', 'OpenWith'
+                //);
+
+                //var sDocumentUrl = "http://192.168.5.220:5005" + path;
+                //var ns = ITHit.WebDAV.Client;
+                //var oNs = ITHit.WebDAV.Client.DocManager;
+                //var session = new ns.WebDavSession();
+                //console.log(session);
+                //if (oNs.IsMicrosoftOfficeDocument(sDocumentUrl)) {
+                    //oNs.MicrosoftOfficeEditDocument(sDocumentUrl, errorCallback());
+                //} else {
+                    //oNs.DavProtocolEditDocument(sDocumentUrl, null, errorCallback());
+                //}
+
                 let href = window.location.href;
                 if (!href.indexOf("#")) {
                     href += "#";
                 }
                 if (href !== event.href) return;
-                path = path + "/" + filename;
-                ITHit.WebDAV.Client.DocManager.EditDocument("http://192.168.5.220:5005" + path, "http://192.168.5.220:5005/", errorCallback);
+
+                ITHit.WebDAV.Client.DocManager.EditDocument("http://192.168.5.220:5005" + filePath, "http://192.168.5.220:5005/", errorCallback);
             }
 
             function errorCallback() {
-                var installerFilePath = "Plugins/" + ITHit.WebDAV.Client.DocManager.GetInstallFileName();
+                //let installerFilePath = "Plugins/" + ITHit.WebDAV.Client.DocManager.GetInstallFileName();
 
-                if (confirm("Opening this type of file requires a protocol installation. Select OK to download the protocol installer.")){
-                    window.open(installerFilePath);
-                }
+                //if (confirm("Opening this type of file requires a protocol installation. Select OK to download the protocol installer.")){
+                    //window.open(installerFilePath);
+                //}
             }
         </script>
     </body>
